refactor: 기존 서비스에서 sqs publisher 를 의존하는 방법에서, 컨트롤러가 서비스와 publisher 를 의존하게 변경, 그에 따른 일부 테스트 생성자 반환타입 수정
