+++ src/main/java/back/ecommerce/service/auth/AuthService.java
@@ -6,17 +6,16 @@ import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import back.ecommerce.auth.token.EmailCodeProvider;
 import back.ecommerce.auth.token.Token;
+import back.ecommerce.common.generator.UuidGenerator;
 import back.ecommerce.auth.token.TokenProvider;
 import back.ecommerce.domain.user.User;
+import back.ecommerce.dto.response.auth.SignUpDto;
 import back.ecommerce.dto.response.auth.TokenResponse;
 import back.ecommerce.dto.response.user.SignUpResponse;
 import back.ecommerce.exception.PasswordNotMatchedException;
 import back.ecommerce.exception.UserNotFoundException;
 import back.ecommerce.repository.user.UserRepository;
-import back.ecommerce.service.auth.email.EmailSender;
-import back.ecommerce.service.auth.email.SignUpEmailMessage;
 import lombok.RequiredArgsConstructor;
 
 @Service
@@ -29,8 +28,7 @@ public class AuthService {
 	private final UserRepository userRepository;
 	private final PasswordEncoder passwordEncoder;
 	private final TokenProvider tokenProvider;
-	private final EmailCodeProvider emailCodeProvider;
-	private final EmailSender emailSender;
+	private final UuidGenerator uuidGenerator;
 	private final SignUpService signUpService;
 
 	@Transactional(readOnly = true)
@@ -44,12 +42,10 @@ public class AuthService {
 		return TokenResponse.create(token.getValue(), token.getExpireTime(), token.getType());
 	}
 
-	public SignUpResponse signUp(String email, String password) {
-		String code = emailCodeProvider.create();
+	public SignUpDto signUp(String email, String password) {
+		String code = uuidGenerator.create();
 		signUpService.cachingSignUpInfo(code, email, password, EMAIL_TOKEN_EXPIRED_TIME);
-		SignUpEmailMessage message = new SignUpEmailMessage(email, code, EMAIL_TOKEN_EXPIRED_TIME);
-		emailSender.send(message);
-		return new SignUpResponse(email, LocalDateTime.now());
+		return new SignUpDto(email, code);
 	}
 
 	public SignUpResponse verifiedEmailCode(String code) {
